-Added two classes (one for importing to the database and another for creating default categories, etc) - Optimized parts of the two classes - Removed the extra class that had become irrelevant - Made some of the codes more readable -Fixed some mistakes

+++ ExpenseTrackerLibrary/ITransaction.cs
@@ -24,7 +24,7 @@ namespace ExpenseTrackerLibrary
         /// <summary>
         /// The amount of money involved in the Transaction.
         /// </summary>
-        float Amount { get; set; }
+        decimal Amount { get; set; }
 
         /// <summary>
         /// The type of transaction that took place; either Expense, Earning, Owed, Debt.
@@ -72,7 +72,6 @@ namespace ExpenseTrackerLibrary
         /// otherwise, there is no note and Note property is null.
         /// </summary>
         bool HasNote { get; }
-
         
     }
 }
